fix: firefox v143 breaks bbcode spoiler styling
Firefox v143 changes the styling of `<details>` elements, forcing the inline horizontal padding to stretch onto the line above and below and add extra "vertical padding". This commit changes the `display` of the `<details>` element used in the bbcode spoiler to `inline-block` which mimics the original behaviour prior to the firefox change.