fix: error 500 on similar page when filter returns 0 results
The similar page assumes that a torrent exists for a torrent, otherwise returning a 404. However, if a user uses the search on the similar page, and uses a search query that returns 0 torrents, the assumption that a torrent exists breaks. In this case, we now fall back to an empty array instead.